# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Description of my API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/create_moderator": {
            "post": {
                "tags": [
                    "Admin Moderators"
                ],
                "summary": "Create a new moderator",
                "description": "Creates a new moderator user with the specified email, name, and password.",
                "operationId": "7e4ebcb9ee70f6386665c0fb133492bc",
                "requestBody": {
                    "description": "Moderator details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Moderator's email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "name": {
                                        "description": "Moderator's name",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Moderator's password (min: 8 characters)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Moderator created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Moderator created successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/update_moderator": {
            "post": {
                "tags": [
                    "Admin Moderators"
                ],
                "summary": "Update moderator information",
                "description": "Updates the information of an existing moderator user.",
                "operationId": "cf4ca746ae6dc311c363f94c37429182",
                "requestBody": {
                    "description": "Moderator details to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the moderator user to update",
                                        "type": "integer"
                                    },
                                    "email": {
                                        "description": "New email for the moderator",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "name": {
                                        "description": "New name for the moderator",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "New password for the moderator (min: 8 characters)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Moderator updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Moderator updated successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/single_page_moderator": {
            "get": {
                "tags": [
                    "Admin Moderators"
                ],
                "summary": "Get information of a single moderator",
                "description": "Retrieves information of a single moderator user based on the provided moderator_id.",
                "operationId": "045ab5753ed5d8ffc5ace782dac13524",
                "parameters": [
                    {
                        "name": "moderator_id",
                        "in": "query",
                        "description": "ID of the moderator user to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moderator information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Moderator not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Not Found Moderator",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/admin/get_all_moderators": {
            "get": {
                "tags": [
                    "Admin Moderators"
                ],
                "summary": "Get all moderators",
                "description": "Retrieves a paginated list of all moderators.",
                "operationId": "56d6af942503bf6422b4eef3e0941c7f",
                "responses": {
                    "200": {
                        "description": "Moderators retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    }
}